+++ src/main/resources/META-INF/plugin.xml
@@ -103,6 +103,9 @@
             <implementation-class>io.acari.DDLC.DDLCHackComponent</implementation-class>
         </component>
         <component>
+            <implementation-class>io.acari.DDLC.DDLCHackComponentJava</implementation-class>
+        </component>
+        <component>
             <implementation-class>io.acari.DDLC.chibi.DDLCChibiComponent</implementation-class>
         </component>
         <component>
